refactor: mili@4 is now the minimum required version
BREAKING CHANGE: bump mili version requirement to 4
