+++ .czrc
@@ -1,1 +1,3 @@
-{ "path": "cz-conventional-changelog" }
+{
+  "path": "cz-conventional-changelog"
+}
