+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -8,6 +8,6 @@ List the issues solved.(e.g. 'close #xxx, close #xxx and close #xxx')
 
 **Please check if the PR fulfills these requirements**
 
-- [ ] Have you followed the guidelines in our [Contributing document](https://github.com/Val-istar-Guo/mili/blob/master/.github/CODE_OF_CONDUCT.md)?
+- [ ] Have you followed the guidelines in our [Contributing document](https://github.com/mili-project-manager/mtpl-nestjs/blob/master/.github/CODE_OF_CONDUCT.md)?
 - [ ] Have you written new tests for your core changes, as applicable?
 - [ ] Have you written or modified docs for your core changes, as applicable?
