+++ package.json
@@ -16,19 +16,23 @@
     "commit": "git-cz",
     "release": "standard-version",
     "release:first": "standard-version --first-release",
-    "postpublish": "git push"
+    "postpublish": "git push",
+    "prepare": "is-ci || husky install",
+    "ci": "git-cz",
+    "release:alpha": "standard-version --prerelease alpha"
   },
   "devDependencies": {
-    "@commitlint/cli": "^9.1.1",
-    "@commitlint/config-conventional": "^9.1.1",
-    "commitizen": "^4.1.2",
-    "cz-conventional-changelog": "^3.2.0",
-    "standard-version": "^8.0.2",
-    "husky": "^4.2.5"
+    "@commitlint/cli": "^12.1.4",
+    "@commitlint/config-conventional": "^12.1.4",
+    "commitizen": "^4.2.4",
+    "cz-conventional-changelog": "^3.3.0",
+    "husky": "^5.2.0",
+    "is-ci": "^3.0.0",
+    "standard-version": "^9.3.0"
   },
   "main": "index.js",
   "dependencies": {
-    "mili": "^3.9.3"
+    "mili": "^4.5.2"
   },
   "repository": {
     "type": "git",
