--- template/.commitlintrc.yml
@@ -1,2 +1,0 @@
-extends:
-  - '@commitlint/config-conventional'
