--- template/.czrc
@@ -1,1 +1,0 @@
-{ "path": "cz-conventional-changelog" }
