--- template/.eslintignore
@@ -1,4 +1,0 @@
-# Dist
-/dist
-/build
-ava.config.js
