--- template/.gitignore.mustache
@@ -1,23 +1,0 @@
-/log
-
-# Mac
-.DS_Store
-
-# Editor
-*.swp
-
-# NodeJS
-node_modules
-npm-debug.log*
-
-# Compile
-dist
-views/index.*.html
-
-# Tmp
-/database
-tworkspace
-src/utils/case-for-run.ts
-packge-lock.json
-
-.env
