--- template/.husky/pre-commit
@@ -1,6 +1,0 @@
-#!/bin/sh
-. "$(dirname "$0")/_/husky.sh"
-
-npm test
-npx --no-install lint-staged
-npx --no-install mili check --diff --no-deps --ignore "**/node_modules"
