--- template/.lintstagedrc.yml
@@ -1,3 +1,0 @@
-'**/*.{ts}':
-  - 'eslint --fix'
-  - 'git add'
