--- template/.vscode/settings.json
@@ -1,4 +1,0 @@
-{
-  "editor.defaultFormatter": "dbaeumer.vscode-eslint",
-  "eslint.validate": ["typescript", "typescriptreact"],
-}
