--- template/README.md.mustache
@@ -1,9 +1,0 @@
-# {{project.name}}
-
-{{#project.repository}}
-[![dependencies](https://img.shields.io/david/{{project.repository.owner}}/{{project.repository.name}}.svg?style=flat-square)](https://www.npmjs.com/package/{{porject.name}})
-{{/project.repository}}
-
-
-<!-- custom -->
-<!-- custom -->
