+++ templates/.gitignore.hbs
