+++ templates/README.md.hbs
@@ -1,0 +1,9 @@
+# {{project.name}}
+
+{{#if repository.isGitHubRepo}}
+[![dependencies](https://img.shields.io/david/{{repository.gitHubReopOwner}}/{{repository.gitHubRepoName}}.svg?style=flat-square)](https://www.npmjs.com/package/{{npm.name}})
+{{/if}}
+
+
+<!-- custom -->
+<!-- custom -->
