+++ templates/ava.config.js
