+++ templates/index.ts
@@ -1,4 +1,3 @@
 import bootstrap from './src/index'
 
 bootstrap()
-
