+++ templates/src/config/index.ts
