+++ templates/src/index.ts
@@ -2,7 +2,6 @@ import { NestFactory } from '@nestjs/core'
 import { SwaggerModule, DocumentBuilder } from '@nestjs/swagger'
 import { AppModule } from './app-module'
 import { ValidationPipe } from '@nestjs/common'
-import { name, version } from '@root/package.json'
 import * as SELF from '@/config/self'
 import '@/config'
 import morgan from 'morgan'
@@ -13,12 +12,16 @@ async function bootstrap(): Promise<void> {
   app.useGlobalPipes(new ValidationPipe())
   app.use(morgan('tiny'))
 
-  const options = new DocumentBuilder()
-    .setTitle(name)
-    .setVersion(version)
-    .addTag('reporter', 'sotest 报告服务')
+   const builder = new DocumentBuilder()
+    .setTitle(SELF.title)
+    .setVersion(SELF.version)
     .addServer(SELF.origin)
-    .build()
+
+  for (const tag of SELF.tags) {
+    builder.addTag(tag.name, tag.description)
+  }
+
+  const options = builder.build()
   const document = SwaggerModule.createDocument(app, options)
   const httpAdapter = app.getHttpAdapter()
   httpAdapter.use('/api/swagger', (req, res) => {
